@@ -1,6 +1,6 @@
 package omsModel;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Table()
